Roy|000|Restrcutured Course content
Aim of this refactoring is to minimise calls to couchdb when loading up a parent
course element.
Course will be saved as a single document with all modules,
module will be saved as a single document with all chapters
and so on an so forth.
However all content will also be saved indivifually,which means though
a course document wll have all modules, all of the modules will also be saved
as individual couch db documents.
This structure will faciliate reuse of child components at the same time ensuring that
only one request is made to couch while loading up a parent.

Change-Id: Ifa450c0e3166abf01329d6fd7476337fb95dfeb2